backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.11.0.post1",
    "numpy>=2.3.2",
    "pandas>=2.3.1",
    "rich>=14.1.0",
    "scikit-learn>=1.7.1",
    "sentence-transformers>=5.0.0",
    "tabulate>=0.9.0",
]
